Initial commit
The RPi_MFD_Interface is a Atmega16-based device to connect the
RaspberryPi with the CarAudio head unit. Its main purpose is to read the
buttons of the car audio and send it to the pi.  The second purpose is
to power the pi with control over booting up and shutting down of the
raspberry pi to prevent it killing the cars battery.